% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate billions of data points into a reduced-representation matrix}
\usage{
aggregate(x = "twk", aggregation = "character",
  reduction = "character", xbins = "numeric", ybins = "numeric",
  minCount = "numeric", n_threads = "numeric", verbose = "logical",
  progress = "logical", ...)
}
\arguments{
\item{x}{Input \code{\link{twk}} class holding the target path.}

\item{aggregation}{String representing the aggregation function to use
on the input data.}

\item{reduction}{String representing the reduction function to use after
after data has been aggregated.}

\item{xbins}{Number of bins (pixels) in the x-dimension.}

\item{ybins}{Number of bins (pixels) in the y-dimension.}

\item{minCount}{Only pixels with at least this many observations are reported.}

\item{verbose}{Flag triggering verbose output (written to std::cerr). This
will usually, but not always, be appropriately handled by R.}

\item{progress}{This flag will trigger a detached thread internal to Tomahawk
that will tick progression every 30 seconds to the console. Note that R
NEVER cleans up this thread and it will tick until the R instance is closed.}

\item{threads}{Number of aggregation/reduction threads to use.}
}
\value{
Returns a \code{\link{twk_agg}} object with the aggregated matrix.
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
# Assuming you have a file called "test.two" in your downloads directory.
twk<-openTomahawkOutput("~/Downloads/test.two")
agg <- aggregateOutput(twk, "r2" ,"count", 1000, 1000, 50, verbose = T, threads = 4)
plot(agg, normalized = TRUE)
plot(agg, normalized = FALSE)
agg
}
\seealso{
\code{\link{twk_data}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
