% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{calculateLDSingle}
\alias{calculateLDSingle}
\title{Calculate pairwise LD for a target SNV to its neighbourhood}
\usage{
calculateLDSingle(x = "twk", interval = "ANY", window = "numeric",
  minP = "numeric", minR2 = "numeric", threads = "numeric",
  verbose = "logical", progress = "logical", ...)
}
\arguments{
\item{x}{Integer representing the sorted state of the file.}

\item{interval}{This parameter can be either a character string 
encoded as "CHR:POS" or a GenomicRanges object overlapping a 
single position.}

\item{window}{Neighbourhood in base-pairs.}

\item{minP}{Largest P-value to report.}

\item{minR2}{Smallest R-squared (R2) value to report.}

\item{threads}{Number of Tomahawk threads used to unpack and compute the
association data.}

\item{verbose}{Flag triggering verbose output (written to std::cerr). This
will usually, but not always, be appropriately handled by R.}

\item{progress}{This flag will trigger a detached thread internal to Tomahawk
that will tick progression every 30 seconds to the console. Note that R
NEVER cleans up this thread and it will tick until the R instance is closed.}
}
\value{
Returns a \code{\link{twk}} object with the loaded data.
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
# This example assumes you have a Tomahawk file called "1kgp3_chr6.twk" in
# your current working directory.
twk2<-new("twk")
twk2@file.path <- "1kgp3_chr6.twk"
y<-calculateLDSingle(twk2, "6:20682622", threads = 4, window = 100000)
# Example using GenomicRanges class.
require(GenomicRanges)
g <- GRanges("6", IRanges(20682622, 20682622))
y<-calculateLDSingle(twk2, g, threads = 4, window = 100000)
}
\seealso{
\code{\link{twk_data}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
