% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{displayColors}
\alias{displayColors}
\title{Displays the color schemes available in rtomahawk}
\usage{
displayColors()
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
displayColors()
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
