% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{import}
\alias{import}
\title{Imports a htslib-compatible variant-call file into Tomahawk}
\usage{
import(input = "character", output = "character",
  missingness = "numeric", block_size = "integer",
  compression_level = "integer", filter_univariate = "logical", ...)
}
\arguments{
\item{input}{Input string pointing to the htslib-compatible file.}

\item{output}{Output string pointing to the Tomahawk file.}

\item{block_size}{Advanced use: number of variants packed in each internal Tomahawk
block. This could have potential impact on the parallel throughput capabilities
of Tomahawk on large cohorts.}

\item{compression_level}{Compressing level in the range [1, 22] for compressing the
Tomahawk file. Larger numbers results in smaller resting file size at the expense
of importing speeds. Decompressing speeds are largely unaffected by the compression 
level.}

\item{filter_univariate}{Logical (boolean) flag set to \code{TRUE} if univariate (monomorphic)
should be filtered out. If this flag is set to \code{FALSE} then monomorphic sites
are left in the file. The recommendation is to remove these as they contribute no
information.}

\item{missnigness}{Numerical value in the range [0, 1] representing the
largest fraction of missing values that are allowed. If the threshold is
violated then the variant site is filtered out.}
}
\value{
Returns a \code{twk} class with the file handle pointing to the newly imported
   Tomahawk file.
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
# This example assumes you have a Tomahawk file called "1kgp3_chr20.bcf" in
# your current working directory.
twk <- import("1kgp3_chr20.bcf","~/Downloads/1kgp3_chr20")
twk
}
\seealso{
\code{\link{twk_data}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
