% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{loadAggregate}
\alias{loadAggregate}
\title{Load aggregation object generated by tomahawk}
\usage{
loadAggregate(input = "character", ...)
}
\arguments{
\item{input}{Input target file path.}
}
\value{
Returns a \code{\link{twk_agg}}.
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
# Assuming you have a file called "agg.twa" in your downloads directory.
agg<-loadAggregate("~/Downloads/agg.twa")
plot(agg, normalize = TRUE)
plot(agg, normalize = FALSE)
agg
}
\seealso{
\code{\link{twk_data}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
