% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{openTomahawkOutput}
\alias{openTomahawkOutput}
\title{Open a file-handle to a Tomahawk object on disk}
\usage{
openTomahawkOutput(input = "character", ...)
}
\arguments{
\item{input}{Input string pointing to the Tomahawk output file.}
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
# This example assumes you have a Tomahawk file called "1kgp3_chr6.two" in
# your current working directory.
f<-setFilters(minR2=0.5)
twk<-openTomahawkOutput("1kgp3_chr6.two")
twk
}
\seealso{
\code{\link{twk_data}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
