% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{plotAggregation}
\alias{plotAggregation}
\title{General plotting method for Tomahawk aggregation.}
\usage{
plotAggregation(x = "twk_agg", normalize = "logical", ...)
}
\arguments{
\item{x}{Instance of a \code{\link{twk}} object. This is the 
computed information that is returned from the underlying 
\code{.Call} to Tomahawk. Note that no data is stored in
this object in this case.}
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
# This example assumes you have a Tomahawk file called "1kgp3_chr6.two" in
# your current working directory.
twk<-rtomahawk::OpenTomahawkOutput("1kgp3_chr6.two")
x<-aggregate(twk,"r2","count",1000,1000,50,verbose=T,threads=4)
plot(x,normalized=TRUE)
plot(x,normalized=FALSE)
x
}
\seealso{
\code{\link{twk_data}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
