% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{plotCytoband}
\alias{plotCytoband}
\title{Plots a cytoband}
\usage{
plotCytoband(chr, release = "hg19", zoom = NULL)
}
\arguments{
\item{chr}{Target chromosome name.}

\item{release}{Target release name. Currently supported values are "hg19" and "hg38".}

\item{zoom}{Tuple (from,to) for a region to highlight.}
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
plotCytoband("chr6", "hg19")
plotCytoband("chr6", "hg19", zoom = c(10e6,20e6))
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
