% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{plotLD}
\alias{plotLD}
\title{Plot LD square}
\usage{
plotLD(x = "twk", from = "ANY", to = "ANY", colors = "ANY",
  bg = "ANY", opacity = "ANY", upper = "logical",
  lower = "logical", annotate = "logical", really = "logical", ...)
}
\arguments{
\item{x}{Input Tomahawk class.}

\item{from}{Start position in base-pairs.}

\item{to}{End position in base-pairs.}

\item{colors}{Vector of colors used to plot with. These are different from \code{col}
as these colors are internally parsed to add opacity.}

\item{opacity}{Logical flag/numerical value controlling opacity levels of colors. If
set to the logical \code{FALSE}, colours will have a uniform opacity of 1. If set
to a numerical value, colours will uniformly share that opacity level. If unset,
opacity will be a gradient from [0.1, 1] starting at R2 > 0.1.}

\item{upper}{Plot only the upper triangular of values.}

\item{lower}{Plot only the lower triangular of values.}

\item{really}{Logical flag to prevent plotting too many points.
Set this to \code{TRUE} at your own peril.}
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
# This example assumes you have a Tomahawk file called "1kgp3_chr6.two" in
# your current working directory.
f<-setFilters(minR2=0.1)
twk<-openTomahawkOutput("1kgp3_chr6.two")
y<-readRecords(twk,f,"6:5e6-10e6",really=TRUE)
plotLD(y, ylim=c(5e6, 10e6), xlim=c(5e6, 10e6))
plotLD(y, ylim=c(5e6, 10e6), xlim=c(5e6, 10e6), upper = TRUE)
plotLD(y, ylim=c(5e6, 10e6), xlim=c(5e6, 10e6), lower = TRUE)
}
\seealso{
\code{\link{twk_data}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
