% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{plotLZ}
\alias{plotLZ}
\title{Plotting and computing target LD}
\usage{
plotLZ(x = "twk", interval = "ANY", snp = "data.frame",
  gmap = "list", window = "numeric", minP = "numeric",
  minR2 = "numeric", threads = "numeric", verbose = "numeric", ...)
}
\arguments{
\item{x}{Integer representing the sorted state of the file.}

\item{interval}{This parameter can be either a character string 
encoded as "CHR:POS" or a GenomicRanges object overlapping a 
single position.}

\item{snp}{A data.frame with the input P-values and their positions
and names. The column names "SNP", "Position", and "p" MUST exist
in this data frame.}

\item{gmap}{List of recombination hotspots for the species in question.
The list entries MUST be data.frames having the columns "position", 
"COMBINED_rate.cM.Mb.", and "Genetic_Map.cM.". This data is provided
for human build hg19 in rtomahawk. It can be loaded as follows:
data(gmap)}

\item{window}{Neighbourhood in base-pairs.}

\item{minP}{Largest P-value to report.}

\item{minR2}{Smallest R-squared (R2) value to report.}

\item{threads}{Number of Tomahawk threads used to unpack and compute the
association data.}

\item{verbose}{Flag triggering verbose output (written to std::cerr). This
will usually, but not always, be appropriately handled by R.}
}
\value{
Returns a \code{\link{twk}} object with the loaded data.
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
data(gmap)
twk2<-new("twk")
twk2@file.path <- "1kgp3_chr6.twk"
plotLZ(twk2, "6:20694884", snp, gmap, window=1000000, minR2=0)
}
\seealso{
\code{\link{twk_data}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
