% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{readRecords}
\alias{readRecords}
\title{Reads Tomahawk output records into memory}
\usage{
readRecords(x = "twk", intervals = "ANY", filters = "ANY",
  really = "logical", ...)
}
\arguments{
\item{x}{Input Tomahawk class.}

\item{intervals}{Filter the records using the provided intervals. The
intervals can be in either character form or as a \code{GenomicRanges}
object.}

\item{filters}{Filter the records using this \code{twk_filter} class.
This parameter can be set to \code{null} to skip all filtering
procedures and read all records as is.}

\item{really}{Logical flag to prevent plotting too many points.
Set this to \code{TRUE} at your own peril.}
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
# This example assumes you have a Tomahawk file called "1kgp3_chr6.two" in
# your current working directory.
f<-setFilters(minR2=0.5)
twk<-openTomahawkOutput("1kgp3_chr6.two")
y<-readRecords(twk,"6:5e6-10e6",f,really=TRUE)
y
# Example using GenomicRanges class.
require(GenomicRanges)
g <- GRanges("6", IRanges(6e6, 7e6))
y<-readRecords(twk,g,f,really=TRUE)
y
}
\seealso{
\code{\link{twk_data}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
