\name{rtomahawk-package}
\alias{rtomahawk-package}
\alias{rtomahawk}
\docType{package}
\title{
  Tools for computing, querying and storing linkage-disequilibrium data
}
\description{
  Tomahawk is a unified framework for computing, storing, querying, aggregating, and plotting genome-wide linkage-disequilibrium (LD) data in population-scaled cohorts. This package is an R interface to Tomahawk
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Marcus D. R. Klarqvist <mk819@cam.ac.uk>

Maintainer: Marcus D. R. Klarqvist <mk819@cam.ac.uk>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}