% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{package}
\name{rtomahawk}
\alias{rtomahawk}
\alias{rtomahawk-package}
\title{rtomahawk}
\description{
Description of your package
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
