% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{setFilters}
\alias{setFilters}
\title{Set filters for a Tomahawk output records}
\usage{
setFilters(x = "ANY", flagInclude = "numeric",
  flagExclude = "numeric", minD = "numeric", maxD = "numeric",
  minDprime = "numeric", maxDprime = "numeric", minR = "numeric",
  maxR = "numeric", minR2 = "numeric", maxR2 = "numeric",
  minP = "numeric", maxP = "numeric", minP1 = "numeric",
  maxP1 = "numeric", minP2 = "numeric", maxP2 = "numeric",
  minQ1 = "numeric", maxQ1 = "numeric", minQ2 = "numeric",
  maxQ2 = "numeric", minChiSqFisher = "numeric",
  maxChiSqFisher = "numeric", minChiSqModel = "numeric",
  maxChiSqModel = "numeric", upperOnly = "logical",
  lowerOnly = "logical", ...)
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\section{Slots}{

\describe{
\item{\code{flagInclude}}{Integer representing FLAGs to include.}

\item{\code{flagExclude}}{Integer representing FLAGs to exclude.}

\item{\code{minD}}{Smallest Lewontin D value.}

\item{\code{maxD}}{Largest Lewontin D value.}

\item{\code{minDprime}}{Smallest scaled Lewontin D value.}

\item{\code{minDprime}}{Largest scald Lewontin D value.}

\item{\code{minR}}{Smallest Pearson's correlation coefficient value.}

\item{\code{minR}}{Largest Pearson's correlation coefficient value.}

\item{\code{minR2}}{Smallest scaled Pearson's correlation coefficient value.}

\item{\code{minR2}}{Largest scaled Pearson's correlation coefficient value.}

\item{\code{minP}}{Smallest Fisher's exact test P-value.}

\item{\code{minP}}{Largest Fisher's exact test P-value.}

\item{\code{minP1}}{Smallest count of REF-REF.}

\item{\code{minP1}}{Largest count of REF-REF.}

\item{\code{minP2}}{Smallest count of REF-ALT.}

\item{\code{minP2}}{Largest count of REF-ALT.}

\item{\code{minQ1}}{Smallest count of ALT-REF.}

\item{\code{minQ1}}{Largest count of ALT-REF.}

\item{\code{minQ2}}{Smallest count of ALT-ALT.}

\item{\code{minQ2}}{Largest count of ALT-ALT.}

\item{\code{minChiSqFisher}}{Smallest Chi-squared test of the 2x2 haplotype table.}

\item{\code{minChiSqFisher}}{Largest Chi-squared test of the 2x2 haplotype table.}

\item{\code{minChiSqModel}}{Smallest Chi-squared test of the 4x4 genotype table.}

\item{\code{minChiSqModel}}{Largest Chi-squared test of the 4x4 genotype table.}

\item{\code{upperOnly}}{Logical flag indicating whether upper triangular values only should be kept.}

\item{\code{lowerOnly}}{Logical flag indicating whether lower triangular values only should be kept.}
}}

\seealso{
\code{\link{twk_index}}, \code{\link{twk_data}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
