% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{tomahawkVersion}
\alias{tomahawkVersion}
\title{Print Tomahawk version string}
\usage{
tomahawkVersion(x = "ANY", ...)
}
\value{
Prints the used version of rtomahawk and the C/C++ 
   shared object versions used by Tomahawk.
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
tomahawkVersion()
}
