% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{class}
\name{twk-class}
\alias{twk-class}
\alias{twk}
\title{Tomahawk output class}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\section{Slots}{

\describe{
\item{\code{file.path}}{Character file path to the target TWO file}

\item{\code{index}}{Instance of \code{\link{twk_index}}}

\item{\code{header}}{Instance of \code{\link{twk_header}}}

\item{\code{data}}{Instance of \code{\link{twk_data}}}
}}

\seealso{
\code{\link{twk_index}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
