% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{class}
\name{twk_agg-class}
\alias{twk_agg-class}
\alias{twk_agg}
\title{Tomahawk output aggregation class}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\section{Slots}{

\describe{
\item{\code{twk}}{Instance of a \code{\link{twk}} object. This is the 
computed information that is returned from the underlying 
\code{.Call} to Tomahawk. Note that no data is stored in
this object in this case.}

\item{\code{n}}{Total number of data points in the reduced output matrix.}

\item{\code{x}}{Number of bins in the X dimension.}

\item{\code{y}}{Number of bins in the Y dimension.}

\item{\code{bpx}}{Base-pairs encompassed in each x-bin.}

\item{\code{bpy}}{Base-pairs encompassed in each y-bin.}

\item{\code{aggregation}}{The function string name used for aggregation.}

\item{\code{reduction}}{The function string name used for reduction.}

\item{\code{n_original}}{Total number of original data points prior to
aggregation.}

\item{\code{range}}{Total number of base-pairs encompassed by the aggregator.}

\item{\code{offsets}}{Data frame representing the cumulative number of base-pairs
used by each chromosome in the file.}

\item{\code{data}}{Output aggregated matrix of size [x, y] such that x*y = n.}
}}

\seealso{
\code{\link{twk_index}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
