% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{class}
\name{twk_data-class}
\alias{twk_data-class}
\alias{twk_data}
\title{Tomahawk output data class}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\section{Slots}{

\describe{
\item{\code{data}}{Main storage for data}

\item{\code{offset}}{Data frame of offsets vectors}
}}

\seealso{
\code{\link{twk_index}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
\author{
Marcus D. R. Klarqvist <\email{mk819@cam.ac.uk}> |
<\href{https://mdrk.me}{https://mdrk.me}>
}
