% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{class}
\name{twk_index-class}
\alias{twk_index-class}
\alias{twk_index}
\title{Tomahawk output index class}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\section{Slots}{

\describe{
\item{\code{state}}{Integer representing the sorted state of the file}

\item{\code{records}}{Data frame of index records corresponding to the boundaries of each TWO block}

\item{\code{meta}}{Data frame of meta index records (collapsed records for each chromosome). This slot will only have values if the input file is sorted.}
}}

\seealso{
\code{\link{twk_data}}, \code{\link{twk_header}}, 
\code{\link{twk_filter}}, and \code{\link{twk}}
}
