% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{tomahawkVersion}
\alias{tomahawkVersion}
\title{Print Tomahawk version string}
\usage{
tomahawkVersion(x = "ANY", ...)
}
\arguments{
\item{x}{Description of \code{x}. The main argument in this
example. Most often has such and such properties.}

\item{...}{Additional argument list that might not ever
be used.}

\item{y}{Description of \code{y}. An argument that is rarely
used by \code{"helloworld"} methods.}
}
\value{
Prints the C/C++ shared object versions used by Tomahawk.
}
\description{
Some additional details about this S4 generic and its methods.
The extra blank line between this section and the title is
critical for roxygen2 to differentiate the title from the
description section.
}
\examples{
tomahawkVersion()
}
